authorizing actions

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -16,6 +16,8 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+        
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -26,6 +28,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if(request()->user()->cannot('create', Task::class)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
 
         return TaskResource::make($task);
@@ -47,6 +53,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if($request->user()->cannot('update', $task)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -57,6 +67,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if(request()->user()->cannot('delete', $task)){
+            abort(403, "This action is unauthorized.");
+        }
+
         $task->delete();
 
         return response()->noContent();
